Make URLSearchParamsIterationSource not snapshot
Previously URLSearchParamsIterationSource iterated over a copy of
URLSearchParams, but it is specified in the standard as iterating over
the live list.

Modify it to reference the original URLSearchParams object instead of
making a copy.

Also add web-platform tests for delete during foreach, and a chromium-
specific test for GC of URLSearchParams during iteration.

Remove failing expectations.

BUG=677322

Change-Id: I8c53fd8dd9863fe1146c5b7849d4f08245b37bc1